chore: refactor conditionals folder structure
